+++ R/counts.R
@@ -91,11 +91,13 @@ counts.data.frame <- function(x, cols, sort = FALSE, ..., .name = "freq") {
     cols <- colnames(x)[cols]
   }
 
-  if (length(cols) > 1) {
-    return(counts_n(x[, cols], sort = sort, name = .name))
-  }
+  out <-
+    if (length(cols) > 1) {
+      counts_n(x[, cols], sort = sort, name = .name)
+    } else {
+      vector2df(counts(x[[cols]], sort = sort), cols, .name %||% "freq")
+    }
 
-  out <- vector2df(counts(x[[cols]], sort = sort), cols, .name %||% "freq")
   remake_df(out, x[, cols, drop = FALSE])
 }
 
@@ -135,17 +137,14 @@ props.data.frame <- function(
     cols <- colnames(x)[cols]
   }
 
-  if (length(cols) > 1) {
-    values <- x[, cols]
-    na_ind <- if (na.rm) which(!stats::complete.cases(values))
-    return(props_n(values, sort = sort, name = .name, na_ind = na_ind))
-  }
-
-  out <- vector2df(
-    props(x[[cols]], sort = sort, na.rm = na.rm),
-    cols,
-    .name %||% "prop"
-  )
+  out <-
+    if (length(cols) > 1) {
+      values <- x[, cols, drop = FALSE]
+      na_ind <- if (na.rm) which(!stats::complete.cases(values))
+      props_n(values, sort = sort, name = .name, na_ind = na_ind)
+    } else {
+      vector2df(props(x[[cols]], sort = sort, na.rm = na.rm), cols, .name %||% "prop")
+    }
 
   remake_df(out, x[, cols, drop = FALSE])
 }
@@ -186,12 +185,14 @@ props_n <- function(
 ) {
   res <- counts_n(x, name %||% "prop", sort = sort)
   n <- ncol(res) # maybe could use `name`?
+
   if (is.null(na_ind)) {
     big_n <- length(res[[n]])
   } else {
     res[[n]][na_ind] <- NA_real_
     big_n <- length(res[[n]][-na_ind])
   }
+
   res[[n]] <- res[[n]] / big_n
   res
 }
@@ -207,18 +208,11 @@ remake_df <- function(new, old) {
     n <- new_cn[i]
     o <- old_cn[m[i]]
 
-    a <- attributes(old[[o]])
-    a <- a[setdiff(names(a), "class")]
-    attributes(new[[n]]) <- a
-
     if (is.factor(old[[o]])) {
-      new[[n]] <- factor(
-        new[[n]],
-        levels = levels(old[[o]]),
-        # ordered = is.ordered(old[[o]])
-      )
+      new[[n]] <- match(new[[n]], levels(old[[o]]))
     }
 
+    attributes(new[[n]]) <- attributes(old[[o]])
     class(new[[n]]) <- class(old[[o]])
   }