+++ R/insert.R
@@ -12,30 +12,31 @@
 #' insert(letters[1:5], c(2, 4), c("X", "Y"))
 #'
 #' @export
-#'
 insert <- function(x, positions, values) {
   stopifnot(!anyNA(positions))
   positions <- as.integer(positions)
 
-  lv <- length(values)
-  lp <- length(positions)
+  nval <- length(values)
+  npos <- length(positions)
 
-  if (lp == 0L) {
+  if (npos == 0L) {
     stop("positions has no length")
   }
 
-  if (lv == 1L) {
-    values <- rep.int(values, lp)
-  } else if (lv == lp) {
+  if (nval == 1L && !is.list(values)) {
+    values <- rep.int(values, npos)
+  } else if (nval == npos) {
     o <- order(positions)
     positions <- positions[o]
     values <- values[o]
   } else {
-    stop("length(values) must be equal to length(positions) or 1")
+    stop("length(values) must be equal to length(positions) or 1", call. = FALSE)
   }
 
+  seqs <- seq_along(positions)
+  positions <- positions + seqs - 1L
   for (i in seq_along(positions)) {
-    x <- append(x, values[i], positions[i] + i - 2)
+    x <- append0(x, values[i], positions[i])
   }
 
   x