+++ R/nas.R
@@ -60,6 +60,17 @@ remove_na.factor <- function(x) {
   struct(out, class(x), levels = lvls[!na_levels])
 }
 
+#' @rdname remove_na
+#' @export
+remove_na.fact <- function(x) {
+  # browser()
+  x <- fact_na(x, remove = TRUE)
+  at <- attributes(x)
+  x <- x[!is.na(x)]
+  attributes(x) <- at
+  x
+}
+
 #' Omit NA values
 #'
 #' @param x A vector of values