+++ tests/testthat/test-utils.R
@@ -20,8 +20,6 @@ test_that("utils", {
 
   res <- remove_class(struct(1, class = "foo"))
   expect_identical(class(res), "numeric")
-
-  expect_identical(append0(1:3, 4L), 1:4)
 })
 
 test_that("check_interactive() works", {
@@ -42,4 +40,9 @@ test_that("check_interactive() works", {
 test_that("has_char() works", {
   expect_identical(has_char(c(NA, "this", "")), c(FALSE, TRUE,  FALSE))
   expect_identical(has_char(c(1, 2, NA))      , c(FALSE, FALSE, FALSE))
+})
+
+test_that("dupe_check() works", {
+  expect_error(dupe_check(c(1, 1, 2, 3)))
+  expect_error(dupe_check(1:4), NA)
 })